requirements.txt
pandas
numpy
requests
beautifulsoup4
lxml
python-dotenv
ipykernel
jupyterlab




